Avoid MSSQL problems
`UNIQUE` relies on having an index created for the column, and the index's maximum column size is 900 bytes, so VARCHAR(max) with it's 2GB column is a bit of a problem. This is where the table creation failed and we got an error: "Column '$columnName' in table '$tableName' is of a type that is invalid for use as a key column in an index."

By replacing `sa.String()` with `sa.String(255)` we now end up with a create statement which makes MS SQL Server very happy as VARCHAR(255) is within the 900 byte limit for most of the cases. (It seems to naively assume single-byte characters)

Or in some other case we just get rid of the unique constraint